mindfulai-final/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# HTTP client (Groq / local LLM calls)
httpx==0.26.0

# AI / ML (Removed for Vercel deployment - too heavy)
# transformers==4.37.0
# torch==2.1.2
# faiss-cpu==1.8.0
# sentence-transformers==3.0.1
numpy==1.26.3

# Env
python-dotenv==1.0.0
